Update serial_data_arduino_receives_py.ino
This is the respective code for turning on/off and led using voice. This code is uploaded onto a arduino. I used arduino uno but other arduino boards like mega should work. The led is connected to pin number 7 of arduino uno. Please note that i had an extrnal led connected to use pin 7. If you do not have the external circuit (comprising of a led and a resistor and jumper wires), then instead of pin 7, use pin 13 of arduino uno.